Split out Image Color built-ins
