+++ mathics/builtin/drawing/__init__.py
@@ -1,14 +1,22 @@
 """
-Graphics, Drawing, and Images
+Graphics and Drawing
 
-Showing something visually can be don in a number of ways:
+Showing something visually can be done in a number of ways:
 
 <ul>
-  <li>Starting with complete images and modifiying them. The 'Image' function is in this category.
-  <li>Use pre-defined 2D or 3D objects like <url>:'Circle': /doc/reference-of-built-in-symbols/drawing-graphics/circle</url> and <url>:'Cuboid': /doc/reference-of-built-in-symbols/graphics-drawing-and-images/three-dimensional-graphics/cuboid/</url> and place them in a coordiate space.
-  <li>Compute the points of the space using a function. This is done using functions like <url>:'Plot': /doc/reference-of-built-in-symbols/graphics-drawing-and-images/plotting-data/plot</url> and <url>:'ListPlot': /doc/reference-of-built-in-symbols/graphics-drawing-and-images/plotting-data/listplot</url>.
+  <li>Starting with complete images and modifiying them using the 'Image' Built-in function.
+  <li>Use pre-defined 2D or 3D objects like <url>
+  :'Circle':
+  /doc/reference-of-built-in-symbols/drawing-graphics/circle</url> and <url>
+  :'Cuboid': /doc/reference-of-built-in-symbols/graphics-drawing-and-images/three-dimensional-graphics/cuboid/</url> \
+  and place them in a coordiate space.
+  <li>Compute the points of the space using a function. This is done using functions like <url>
+  :'Plot':
+  /doc/reference-of-built-in-symbols/graphics-drawing-and-images/plotting-data/plot</url> \
+  and <url>
+  :'ListPlot': /doc/reference-of-built-in-symbols/graphics-drawing-and-images/plotting-data/listplot</url>.
 </ul>
 """
 
 # This tells documentation how to sort this module
-sort_order = "mathics.builtin.graphing-drawing-and-images"
+sort_order = "mathics.builtin.graphing-and-drawing"
