+++ mathics/core/systemsymbols.py
@@ -119,6 +119,7 @@ SymbolMachinePrecision = Symbol("System`MachinePrecision")
 SymbolMakeBoxes = Symbol("System`MakeBoxes")
 SymbolMap = Symbol("System`Map")
 SymbolMatchQ = Symbol("System`MatchQ")
+SymbolMatrixQ = Symbol("System`MatrixQ")
 SymbolMathMLForm = Symbol("System`MathMLForm")
 SymbolMatrixPower = Symbol("System`MatrixPower")
 SymbolMax = Symbol("System`Max")
@@ -204,6 +205,7 @@ SymbolTan = Symbol("System`Tan")
 SymbolTanh = Symbol("System`Tanh")
 SymbolTeXForm = Symbol("System`TeXForm")
 SymbolThrow = Symbol("System`Throw")
+SymbolThreshold = Symbol("System`Threshold")
 SymbolToString = Symbol("System`ToString")
 SymbolTotal = Symbol("System`Total")
 SymbolTraditionalForm = Symbol("System`TraditionalForm")
